Use codejail to sandbox execution of queries Queries will respect the edx-platform configured resource limits for codejail processes. Both the instructor and student queries will be executed in a sandbox.
