+++ requirements/base.txt
@@ -4,11 +4,12 @@
 #
 #    make upgrade
 #
+-e git+https://github.com/edx/codejail.git@3.0.0#egg=codejail  # via -c requirements/constraints.txt, -r requirements/base.in
 appdirs==1.4.4            # via fs
-django==2.2.12            # via -c requirements/constraints.txt, -r requirements/base.in
+django==2.2.14            # via -c requirements/constraints.txt, -r requirements/base.in
 fs==2.4.11                # via xblock
-lxml==4.5.1               # via xblock
-mako==1.1.2               # via xblock-utils
+lxml==4.5.2               # via xblock
+mako==1.1.3               # via xblock-utils
 markupsafe==1.1.1         # via mako, xblock
 python-dateutil==2.8.1    # via xblock
 pytz==2020.1              # via django, fs, xblock
@@ -16,7 +17,6 @@ pyyaml==5.3.1             # via xblock
 simplejson==3.17.0        # via xblock-utils
 six==1.15.0               # via fs, python-dateutil, xblock
 sqlparse==0.3.1           # via django
-typing==3.7.4.1           # via fs
 web-fragments==0.3.2      # via xblock, xblock-utils
 webob==1.8.6              # via xblock
 xblock-utils==2.1.1       # via -r requirements/base.in
