+++ requirements/quality.txt
@@ -4,32 +4,31 @@
 #
 #    make upgrade
 #
+-e git+https://github.com/edx/codejail.git@3.0.0#egg=codejail  # via -c requirements/constraints.txt, -r requirements/base.txt
 appdirs==1.4.4            # via -r requirements/base.txt, fs
-astroid==2.4.1            # via pylint
-coverage==5.1             # via -r requirements/test.txt
-django==2.2.12            # via -c requirements/constraints.txt, -r requirements/base.txt
+astroid==2.4.2            # via pylint
+coverage==5.2             # via -r requirements/test.txt
+django==2.2.14            # via -c requirements/constraints.txt, -r requirements/base.txt
 edx-opaque-keys==2.1.0    # via -r requirements/test.txt
 fs==2.4.11                # via -r requirements/base.txt, xblock
 isort==4.3.21             # via pylint
 lazy-object-proxy==1.4.3  # via astroid
-lxml==4.5.1               # via -r requirements/base.txt, xblock
-mako==1.1.2               # via -r requirements/base.txt, xblock-utils
+lxml==4.5.2               # via -r requirements/base.txt, xblock
+mako==1.1.3               # via -r requirements/base.txt, xblock-utils
 markupsafe==1.1.1         # via -r requirements/base.txt, mako, xblock
 mccabe==0.6.1             # via pylint
 pbr==5.4.5                # via -r requirements/test.txt, stevedore
 pycodestyle==2.6.0        # via -r requirements/quality.in
-pylint==2.5.2             # via -r requirements/quality.in
+pylint==2.5.3             # via -r requirements/quality.in
 pymongo==3.10.1           # via -r requirements/test.txt, edx-opaque-keys
 python-dateutil==2.8.1    # via -r requirements/base.txt, xblock
 pytz==2020.1              # via -r requirements/base.txt, django, fs, xblock
 pyyaml==5.3.1             # via -r requirements/base.txt, xblock
 simplejson==3.17.0        # via -r requirements/base.txt, xblock-utils
-six==1.15.0               # via -r requirements/base.txt, -r requirements/test.txt, astroid, edx-opaque-keys, fs, python-dateutil, stevedore, xblock
+six==1.15.0               # via -r requirements/base.txt, -r requirements/test.txt, astroid, edx-opaque-keys, fs, python-dateutil, xblock
 sqlparse==0.3.1           # via -r requirements/base.txt, django
 stevedore==1.32.0         # via -r requirements/test.txt, edx-opaque-keys
 toml==0.10.1              # via pylint
-typed-ast==1.4.1          # via astroid
-typing==3.7.4.1           # via -r requirements/base.txt, fs
 web-fragments==0.3.2      # via -r requirements/base.txt, xblock, xblock-utils
 webob==1.8.6              # via -r requirements/base.txt, xblock
 wrapt==1.12.1             # via astroid
