+++ requirements/test.txt
@@ -4,9 +4,9 @@
 #
 #    make upgrade
 #
-coverage==5.1             # via -r requirements/test.in
+coverage==5.2             # via -r requirements/test.in
 edx-opaque-keys==2.1.0    # via -r requirements/test.in
 pbr==5.4.5                # via stevedore
 pymongo==3.10.1           # via edx-opaque-keys
-six==1.15.0               # via edx-opaque-keys, stevedore
+six==1.15.0               # via edx-opaque-keys
 stevedore==1.32.0         # via edx-opaque-keys
