+++ requirements/tox.txt
@@ -5,16 +5,13 @@
 #    make upgrade
 #
 appdirs==1.4.4            # via virtualenv
-distlib==0.3.0            # via virtualenv
+distlib==0.3.1            # via virtualenv
 filelock==3.0.12          # via tox, virtualenv
-importlib-metadata==1.6.0  # via importlib-resources, pluggy, tox, virtualenv
-importlib-resources==1.5.0  # via virtualenv
 packaging==20.4           # via tox
 pluggy==0.13.1            # via tox
-py==1.8.1                 # via tox
+py==1.9.0                 # via tox
 pyparsing==2.4.7          # via packaging
 six==1.15.0               # via packaging, tox, virtualenv
 toml==0.10.1              # via tox
-tox==3.15.1               # via -r requirements/tox.in
-virtualenv==20.0.21       # via tox
-zipp==1.2.0               # via importlib-metadata, importlib-resources
+tox==3.16.1               # via -r requirements/tox.in
+virtualenv==20.0.26       # via tox
