+++ requirements/travis.txt
@@ -5,23 +5,20 @@
 #    make upgrade
 #
 appdirs==1.4.4            # via -r requirements/tox.txt, virtualenv
-certifi==2020.4.5.1       # via requests
+certifi==2020.6.20        # via requests
 chardet==3.0.4            # via requests
-codecov==2.1.3            # via -r requirements/travis.in
-coverage==5.1             # via codecov
-distlib==0.3.0            # via -r requirements/tox.txt, virtualenv
+codecov==2.1.7            # via -r requirements/travis.in
+coverage==5.2             # via codecov
+distlib==0.3.1            # via -r requirements/tox.txt, virtualenv
 filelock==3.0.12          # via -r requirements/tox.txt, tox, virtualenv
-idna==2.9                 # via requests
-importlib-metadata==1.6.0  # via -r requirements/tox.txt, importlib-resources, pluggy, tox, virtualenv
-importlib-resources==1.5.0  # via -r requirements/tox.txt, virtualenv
+idna==2.10                # via requests
 packaging==20.4           # via -r requirements/tox.txt, tox
 pluggy==0.13.1            # via -r requirements/tox.txt, tox
-py==1.8.1                 # via -r requirements/tox.txt, tox
+py==1.9.0                 # via -r requirements/tox.txt, tox
 pyparsing==2.4.7          # via -r requirements/tox.txt, packaging
-requests==2.23.0          # via codecov
+requests==2.24.0          # via codecov
 six==1.15.0               # via -r requirements/tox.txt, packaging, tox, virtualenv
 toml==0.10.1              # via -r requirements/tox.txt, tox
-tox==3.15.1               # via -r requirements/tox.txt
+tox==3.16.1               # via -r requirements/tox.txt
 urllib3==1.25.9           # via requests
-virtualenv==20.0.21       # via -r requirements/tox.txt, tox
-zipp==1.2.0               # via -r requirements/tox.txt, importlib-metadata, importlib-resources
+virtualenv==20.0.26       # via -r requirements/tox.txt, tox
