+++ sql_grader/mixins/grading.py
@@ -1,21 +1,77 @@
 """
 Mixin grading-related functionality
 """
+import logging
+import os.path
+
+from django.utils.translation import ugettext_lazy as _
+
+from codejail.safe_exec import safe_exec, SafeExecException
+
+from xblock.fields import Boolean
 from xblock.fields import Float
 from xblock.fields import Integer
 from xblock.fields import Scope
+from xblock.fields import String
+
 from xblock.scorable import ScorableXBlockMixin
 from xblock.scorable import Score
 
-from ..models import SqlProblem
-from ..models import DATABASES
+from ..problem import all_datasets
 
 
-def _(text):
+log = logging.getLogger('sql_grader')
+
+
+# pylint: disable=unused-argument
+def attempt_safe(dataset, answer_query, verify_query, is_ordered, query):
     """
-    Mock translation for scraping
+    Attempt a SqlProblem, using codejail to sandbox the execution.
     """
-    return text
+    results = {
+        'answer_query': answer_query,
+        'dataset': dataset,
+        'verify_query': verify_query,
+        'is_ordered': is_ordered,
+        'query': query
+    }
+    code = """
+from sql_grader.problem import SqlProblem
+submission_result, answer_result, error, comparison = SqlProblem(
+    answer_query=answer_query,
+    dataset=dataset,
+    verify_query=verify_query,
+    is_ordered=is_ordered
+).attempt(query)
+
+"""
+    # example from edx-platform's use of codejail:
+    # https://github.com/edx/edx-platform/blob/master/common/lib/capa/capa/capa_problem.py#L887
+    # we have to include the path to the entire sql_grader package.
+    python_path = [os.path.abspath(
+        os.path.join(
+            os.path.dirname(__file__),
+            '../..'
+        )
+    )]
+
+    try:
+        safe_exec(code, results, python_path=python_path, slug='sql_grader')
+    except SafeExecException:
+        log.exception(query)
+        # how should resource limits be communicated to the user?
+        results = {
+            'submission_result': None,
+            'answer_result': None,
+            'error': _("We could not execute your query; please try again."),
+            'comparison': None,
+        }
+    return (
+        results['submission_result'],
+        results['answer_result'],
+        results['error'],
+        results['comparison'],
+    )
 
 
 class Scorable(ScorableXBlockMixin):
@@ -82,15 +138,13 @@ class Scorable(ScorableXBlockMixin):
         """
         raw_possible = 1.0
         raw_earned = 0.0
-        database = DATABASES[self.dataset]
-        problem = SqlProblem(
-            database,
+        actual, expected, error, comparison = attempt_safe(
+            self.dataset,
             self.answer_query,
             self.verify_query,
             self.is_ordered,
+            self.raw_response
         )
-        response = self.raw_response
-        actual, expected, error, comparison = problem.attempt(response)
         if comparison:
             raw_earned = 1.0
         score = Score(
@@ -98,3 +152,90 @@ class Scorable(ScorableXBlockMixin):
             raw_possible=raw_possible,
         )
         return (score, actual, expected, error, comparison)
+
+
+class XBlockDataMixin:
+    """
+    Mixin XBlock field data
+    """
+    # pylint: disable=too-few-public-methods
+
+    display_name = String(
+        display_name=_('Display Name'),
+        help=_('The display name for this component.'),
+        default=_('SQL Problem'),
+        scope=Scope.content,
+    )
+    dataset = String(
+        display_name=_('Dataset'),
+        help=_('Which initial dataset/database to be used for queries'),
+        default='rating',
+        scope=Scope.content,
+        values=list(all_datasets()),
+    )
+    answer_query = String(
+        display_name=_('Answer Query'),
+        help=_('A correct response SQL query'),
+        default='',
+        scope=Scope.content,
+        multiline_editor=True,
+    )
+    verify_query = String(
+        display_name=_('Verify Query'),
+        help=_(
+            'A secondary verification SQL query, to be used if the '
+            'answer_query modifies the database (UPDATE, INSERT, DELETE, etc.)'
+        ),
+        default='',
+        scope=Scope.content,
+        multiline_editor=True,
+    )
+    is_ordered = Boolean(
+        display_name=_('Is Ordered?'),
+        help=_('Should results be in order?'),
+        default=False,
+        scope=Scope.content,
+    )
+    editable_fields = [
+        'answer_query',
+        'dataset',
+        'display_name',
+        'verify_query',
+        'is_ordered',
+        'prompt',
+        'weight',
+    ]
+    prompt = String(
+        display_name=_('Prompt'),
+        help=_('Explanatory text to accompany the problem'),
+        default='',
+        scope=Scope.content,
+    )
+    raw_response = String(
+        display_name=_('Submission Query'),
+        help=_('A Submission Query'),
+        default='',
+        scope=Scope.user_state,
+    )
+
+    def provide_context(self, context):  # pragma: no cover
+        """
+        Build a context dictionary to render the student view
+        """
+        context = context or {}
+        context = dict(context)
+        error_class = ''
+        if not bool(self.score) and bool(self.raw_response):
+            error_class = 'error'
+        context.update({
+            'display_name': self.display_name,
+            'prompt': self.prompt,
+            'answer': self.raw_response,
+            'score': self.score,
+            'score_weighted': int(self.score * self.weight),
+            'max_score': int(self.max_score()),
+            'error_class': error_class,
+            'raw_response': self.raw_response,
+            'verify_query': self.verify_query,
+        })
+        return context
