+++ tox.ini
@@ -9,6 +9,7 @@ envlist =
 [testenv]
 deps =
     -rrequirements/test.txt
+    -e git+https://github.com/edx/codejail.git@3.0.0#egg=codejail
     django22: Django>=2.2,<2.3
     django30: Django>=3.0,<3.1
 commands =
